# api/api.yaml
...
openapi: "3.0.0"
info:
  version: 1.0.0
  title: TorrPlay API
  description: A torrent streaming application with memory-managed piece storage.
  termsOfService: https://example.com/terms/
  contact:
    name: TorrPlay Team
    email: support@example.com
    url: https://github.com/torrplay
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
security:
  - basicAuth: []
  - bearerAuth: []
...
paths:
...
  /api/v1/settings:
    get:
      tags:
        - v1
      summary: Get application settings.
      description: Retrieve current application configuration and settings.
      operationId: getSettings
      responses:
        "200":
          description: Settings retrieved successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"

    patch:
      tags:
        - v1
      summary: Update application settings.
      description: Update application configuration. Some settings may require a restart to take effect.
      operationId: updateSettings
      requestBody:
        description: New settings values.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Settings"
      responses:
        "204":
          description: Settings updated successfully.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/torrents:
    get:
      tags:
        - v1
      summary: List all torrents.
      description: Returns a paginated list of all torrents currently managed by the system. Supports filtering by torrent IDs.
      operationId: listTorrents
      parameters:
        - $ref: "#/components/parameters/TorrentCategories"
        - $ref: "#/components/parameters/TorrentHashes"
        - $ref: "#/components/parameters/TorrentNames"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: A paginated list of torrents.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListTorrents"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      tags:
        - v1
      summary: Add a new torrent.
      description: Add a torrent to the database. Supports magnet links, info hashes and torrent files.
      operationId: addTorrent
      requestBody:
        description: |
          Torrent data to add. Choose one of:
          - JSON with magnet link or info hash
          - Multipart form with torrent file and optional metadata
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TorrentAdd"
            examples:
              magnet:
                summary: Add via magnet link
                value:
                  magnet: "magnet:?xt=urn:btih:08ada5a7a6183aae1e09d831df6748d566095a10&dn=Sintel"
                  poster: "https://example.com/poster.jpg"
                  storage: "memory"
                  title: "Sintel"
              magnet_with_data_uri:
                summary: Add via magnet with data URI poster
                value:
                  magnet: "magnet:?xt=urn:btih:dd8255ecdc7ca55fb0bbf81323d87062db1f6d1c&dn=Big+Buck+Bunny"
                  poster: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAQAAAC1HAwCAAAAC0lEQVR42mNkYAAAAAYAAjCB0C8AAAAASUVORK5CYII="
                  title: "Big Buck Bunny"
              hash:
                summary: Add via info hash only
                value:
                  hash: "08ada5a7a6183aae1e09d831df6748d566095a10"
                  storage: "memory"
                  title: "Sintel"
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/TorrentAddFile"
          application/octet-stream:
            schema:
              type: string
              format: binary
            description: Raw torrent file content (filename should be in Content-Disposition header).
      responses:
        "201":
          description: Torrent added successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Torrent"
          headers:
            Location:
              description: URL of the newly created torrent.
              schema:
                type: string
                format: uri
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
...
  /api/v1/torrents/{hash}:
    get:
      tags:
        - v1
      summary: Get torrent metadata.
      description: Retrieve detailed information about a specific torrent by its info hash.
      operationId: getTorrent
      parameters:
        - $ref: "#/components/parameters/Hash"
      responses:
        "200":
          description: Torrent retrieved successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Torrent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "504":
          $ref: "#/components/responses/GatewayTimeout"

    patch:
      tags:
        - v1
      summary: Update torrent metadata.
      description: Update metadata for an existing torrent. Only modifiable fields can be updated.
      operationId: updateTorrent
      parameters:
        - $ref: "#/components/parameters/Hash"
      requestBody:
        description: Torrent metadata to update.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TorrentUpdate"
      responses:
        "200":
          description: Torrent updated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Torrent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      tags:
        - v1
      summary: Delete a torrent.
      description: Remove a torrent from the system. This will also free any memory used by the torrent's pieces.
      operationId: deleteTorrent
      parameters:
        - $ref: "#/components/parameters/Hash"
      responses:
        "204":
          description: Torrent deleted successfully.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /cache:
    post:
      tags:
        - TorrServer
      summary: Get cache statistics (TorrServer compatible).
      description: Compatibility endpoint for TorrServer cache statistics. Returns cache status for a specific torrent.
      operationId: tSCache
      requestBody:
        description: Cache statistics request.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TSCacheRequest"
      responses:
        "200":
          description: Cache statistics retrieved successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TSCacheResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /settings:
    post:
      tags:
        - TorrServer
      summary: Update settings (TorrServer compatible).
      description: Compatibility endpoint for TorrServer settings updates.
      operationId: tSSettings
      requestBody:
        description: TorrServer settings.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TSSettings"
      responses:
        "200":
          description: Settings updated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /stream:
    get:
      tags:
        - TorrServer
      summary: Stream or preload torrent (TorrServer compatible).
      description: Compatibility endpoint for streaming or preloading torrent content with various options.
      operationId: tSStream
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/TSLink"
        - $ref: "#/components/parameters/TSIndexQuery"
        - $ref: "#/components/parameters/TSPlay"
        - $ref: "#/components/parameters/TSPreload"
        - $ref: "#/components/parameters/TSStat"
      responses:
        "200":
          description: Stream or statistics returned successfully.
          content:
            video/*:
              schema:
                type: string
                format: binary
            audio/*:
              schema:
                type: string
                format: binary
            application/json:
              schema:
                $ref: "#/components/schemas/TSTorrentResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
  /torrents:
    post:
      tags:
        - TorrServer
      summary: Manage torrents (TorrServer compatible).
      description: Compatibility endpoint for TorrServer torrent management (add, remove, list).
      operationId: tSTorrents
      requestBody:
        description: Torrent management request.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TSTorrentRequest"
      responses:
        "200":
          description: Operation completed successfully.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/TSTorrentResponse"
                  - type: array
                    items:
                      $ref: "#/components/schemas/TSTorrentResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
  /torrent/upload:
    post:
      tags:
        - TorrServer
      summary: Add a new torrent (TorrServer compatible).
      description: Compatibility endpoint for adding a torrent file.
      operationId: tSTorrentUpload
      requestBody:
        description: |
          - Multipart form with torrent file
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/TorrentAddFile"
      responses:
        "201":
          description: Torrent added successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TSTorrentResponse"
          headers:
            Location:
              description: URL of the newly created torrent.
              schema:
                type: string
                format: uri
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
  /viewed:
    post:
      tags:
        - TorrServer
      summary: Manage torrent view tags (TorrServer compatible).
      description: Compatibility endpoint for TorrServer view tags.
      operationId: tSViewed
      requestBody:
        description: Request to set, list, or remove a view tag.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TSViewedRequest"
      responses:
        "200":
          description: Operation completed successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TSViewedResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  parameters:
...
    TorrentCategories:
      name: categories
      in: query
      description: Comma-separated list of torrent categories to filter results.
      required: false
      style: form
      explode: false
      schema:
        type: array
        items:
          type: string
        example: ["Movies"]
    TorrentHashes:
      name: hashes
      in: query
      description: Comma-separated list of torrent info hashes to filter results.
      required: false
      style: form
      explode: false
      schema:
        type: array
        items:
          type: string
          pattern: "^[0-9a-fA-F]{40}$"
          x-go-type: metainfo.Hash
          x-go-type-import:
            path: github.com/anacrolix/torrent/metainfo
        example: ["08ada5a7a6183aae1e09d831df6748d566095a10"]
    TorrentNames:
      name: names
      in: query
      description: Comma-separated list of torrent names to filter results. The filter performs case-insensitive matching against torrent names.
      required: false
      style: form
      explode: false
      schema:
        type: array
        items:
          type: string
        example: ["Sintel"]
    Limit:
      name: limit
      in: query
      description: Maximum number of torrents to return.
      required: false
      schema:
        type: integer
        minimum: 0
        example: 50
    Offset:
      name: offset
      in: query
      description: Number of torrents to skip for pagination.
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
...
    TSLink:
      name: link
      in: query
      description: Magnet link or info hash of the torrent.
      required: true
      schema:
        type: string
        example: "magnet:?xt=urn:btih:08ada5a7a6183aae1e09d831df6748d566095a10&dn=Sintel"
...
  responses:
...
    BadRequest:
      description: The request was invalid or cannot be served.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Invalid info hash format."
            code: 400
            details: "Info hash must be 40 hex characters."
    Unauthorized:
      description: Authentication failed or is missing.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Unauthorized."
            code: 401
            details: "Authentication token is missing or invalid."
...
    InternalServerError:
      description: An internal server error occurred.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Internal server error."
            code: 500
            details: "Failed to allocate memory for piece storage."
...
  schemas:
...
    Settings:
      type: object
      description: Application configuration settings.
      properties:
        auth:
          $ref: "#/components/schemas/Auth"
        disable_ipv6:
          type: boolean
          description: Disable IPv6 for torrent client.
          example: false
        enable_dlna:
          type: boolean
          description: Enable DLNA media server functionality.
          example: false
        enable_downloader:
          type: boolean
          description: Enable background downloading for torrents with file storage.
          example: false
        file_storage_path:
          type: string
          description: Path to the file storage.
          example: "/torrplay"
        friendly_name:
          type: string
          description: DLNA friendly name displayed to clients.
          minLength: 3
          maxLength: 15
          example: "TorrPlay"
        http_server_port:
          type: integer
          description: TCP port for HTTP server.
          minimum: 1
          maximum: 65535
          example: 8090
        log_format:
          type: string
          description: Log format.
          enum:
            - "json"
            - "text"
          example: "text"
        log_level:
          type: string
          x-go-type: slog.Level
          x-go-type-import:
            path: log/slog
          description: Log level.
          example: "INFO"
        log_store_size:
          type: integer
          description: Number of log entries to store in memory.
          minimum: 0
          maximum: 1000
          example: 100
        max_memory:
          type: integer
          format: int64
          description: Maximum memory allocation for piece storage in bytes.
          minimum: 33554432
          maximum: 2147483648
          example: 67108864
        readahead_percentage:
          type: integer
          description: Readahead percentage value (50-100).
          minimum: 50
          maximum: 100
          example: 90
...